pyproject.toml
[project]
name = "xml-tooltest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.19",
    "langchain-openai>=0.3.6",
    "openai>=1.63.2",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.3.18",
]
